typo refix & videoUrls.ts rename as videoGalleryConfig.ts
