+++ src/components/VideoGallery.tsx
@@ -1,19 +1,11 @@
 import { useState } from "preact/hooks"
-import { videos } from "./videosUrls"
+import { videos } from "./videosGalleryConfig"
 import { StepBack, StepForward } from 'lucide-preact'
-import getPlusGrosseUniteEnFrancais from './joda'
+import getPlusGrosseUniteEnFrancais from './getPlusGrosseUniteEnFrancais'
 import {
-  ZonedDateTime,
-  LocalTime,
-  ZoneId,
-  DateTimeFormatter,
-  LocalDate,
-  LocalDateTime,
-  Period,
-  convert, 
-  Instant, 
+  LocalDate, 
+  Period,  
 } from '@js-joda/core';
-import { Interval } from '@js-joda/extra';
 import '@js-joda/timezone';
 
 const paginationItemsNumber = 5 
@@ -23,32 +15,16 @@ export default function VideoGallery() {
   
   videos.map( (item, index) => {
     const dateStr = item.date.replaceAll('/','-').split("-").reverse().join("-")
-    console.log(dateStr);
     const currentDate = new Date().toJSON().slice(0, 10)
-    console.log(currentDate)
+
     const elapsed = Period.between(
       LocalDate.parse(dateStr),
       LocalDate.parse(currentDate)
     );
     const {plusGrosseUniteEnFrancais, elapsedSplitted} = getPlusGrosseUniteEnFrancais(elapsed)
-    console.log(`Il y a [${elapsedSplitted[0]}] [${plusGrosseUniteEnFrancais}] `)
     videos[index].elapsed = `${elapsedSplitted[0]} ${plusGrosseUniteEnFrancais}`
   })
-    
-    /*
-    const elapsedTest = (now - newDate) / 1000 / 3600 / 24
-    if (elapsedTest < 0) {
-      videos[index].elapsed = Math.floor(elapsedTest) + " heures"
-    }
-    else if (elapsedTest < 31) {
-      videos[index].elapsed = Math.floor(elapsedTest) + " jours"
-    }
-    else if (elapsedTest < 365) {
-      videos[index].elapsed = Math.floor(elapsedTest/30.5) + " mois"
-    }
-    console.log(item.date, " => ", videos[index].elapsed)
-  })
-  */
+
   const videoList = videos.slice(0,-1).slice(pagination * paginationItemsNumber, (pagination * paginationItemsNumber) + paginationItemsNumber)
   const lastVideo = videos.slice(-1)[0]
   
@@ -75,7 +51,7 @@ export default function VideoGallery() {
           </iframe>
           <div class="text-xs flex flex-raw mb-12">
             <div class="mx-8">{playingVideo.title}</div>
-            <div class="mx-8">il y as {playingVideo.elapsed}</div>
+            <div class="mx-8">( il y as {playingVideo.elapsed} )</div>
           </div>
         </div>
         {/* pagination navigation */}
@@ -121,7 +97,7 @@ export default function VideoGallery() {
                 </iframe>
                 <div class="text-xs w-max-[300px] min-w-[300px] mt-2">
                   <div class="mx-2 text-xs w-[250px] whitespace-pre-wrap break-words" >{item.title}</div>
-                  <div class="mx-2">il y as {videoList[index].elapsed}</div>
+                  <div class="mx-2">( il y as {videoList[index].elapsed} )</div>
                 </div>
               </div>
             )
