+++ src/components/videosGalleryConfig.ts
